+++ playwright.dpop-endpoints.config.ts
@@ -1,25 +1,20 @@
+/**
+ * Playwright config for DPoP endpoint tests.
+ */
+
 import { defineConfig, devices } from '@playwright/test';
 
 export default defineConfig({
   testDir: './e2e',
-  testIgnore: ['**/dpop-endpoints.test.ts'],
-  /* Run tests in files in parallel */
-  fullyParallel: true,
-  /* Fail the build on CI if you accidentally left test.only in the source code. */
+  testMatch: '**/dpop-endpoints.test.ts',
+  fullyParallel: false,
   forbidOnly: !!process.env.CI,
-  /* Retry on CI only */
   retries: process.env.CI ? 2 : 0,
-  /* Run tests in a single worker */
   workers: 1,
-  /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
-    /* Base URL to use in actions like `await page.goto('/')`. */
     baseURL: `http://localhost:${process.env.PORT}`,
-
-    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     screenshot: 'on',
   },
-
   /* Configure projects for major browsers */
   projects: [
     {
@@ -37,8 +32,6 @@ export default defineConfig({
       use: { ...devices['Desktop Safari'] },
     },
   ],
-
-  /* Run your local dev server before starting the tests */
   webServer: {
     command: `${process.env.SERVER_COMMAND}`,
     url: `http://localhost:${process.env.PORT}`,